support local testing for mongo test
