+++ smoke-tests-otel-starter/spring-boot-3/src/test/java/io/opentelemetry/spring/smoketest/JvmMongodbSpringStarterSmokeTest.java
@@ -1,0 +1,17 @@
+package io.opentelemetry.smoketest;
+
+import org.junit.jupiter.api.condition.DisabledInNativeImage;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.testcontainers.containers.MongoDBContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@Testcontainers
+@DisabledInNativeImage
+public class JvmMongodbSpringStarterSmokeTest extends AbstractMongodbSpringStarterSmokeTest {
+
+  @Container
+  @ServiceConnection
+  static MongoDBContainer container = new MongoDBContainer("mongo:latest");
+
+}
